+++ frontend/src/queries/useChartBlockMetrics.ts
@@ -14,6 +14,7 @@ const BlockMetricsQuery = gql<BlockMetricsQueryResponse>`
 			avgBlockTime
 			lastTotalMicroCcd
 			lastTotalMicroCcdReleased
+			lastTotalMicroCcdUnlocked
 			avgFinalizationTime
 			lastTotalMicroCcdStaked
 			buckets {